@@ -18,63 +18,61 @@
   },
   "dependencies": {
     "@analytics/google-analytics": "^0.5.3",
-    "@analytics/google-tag-manager": "^0.5.1",
-    "@emotion-icons/bootstrap": "^3.14.0",
+    "@analytics/google-tag-manager": "^0.5.2",
+    "@emotion-icons/bootstrap": "^3.27.0",
     "@emotion/core": "^11.0.0",
-    "@emotion/css": "^11.1.3",
-    "@emotion/react": "^11.4.0",
-    "@emotion/styled": "^11.3.0",
-    "@iconify-icons/bi": "^1.1.5",
-    "@iconify/react": "^1.1.4",
-    "@playwright/test": "^1.16.3",
-    "@restart/ui": "^1.0.1",
-    "@sentry/react": "^6.13.3",
-    "@sentry/tracing": "^6.13.3",
-    "@types/bootstrap": "^5.1.6",
-    "@types/faker": "^5.5.7",
+    "@emotion/css": "^11.7.1",
+    "@emotion/react": "^11.8.2",
+    "@emotion/styled": "^11.8.1",
+    "@iconify-icons/bi": "^1.2.2",
+    "@iconify/react": "^3.2.1",
+    "@playwright/test": "^1.20.2",
+    "@restart/ui": "^1.2.0",
+    "@sentry/react": "^6.19.4",
+    "@sentry/tracing": "^6.19.4",
+    "@types/bootstrap": "^5.1.9",
     "@types/loadjs": "^4.0.1",
-    "@types/lodash-es": "^4.17.5",
+    "@types/lodash-es": "^4.17.6",
     "@types/qs": "^6.9.7",
-    "@types/react-flatpickr": "^3.8.4",
-    "@types/react-router-dom": "^5.1.7",
-    "@typescript-eslint/eslint-plugin": "^4.28.0",
-    "@typescript-eslint/parser": "^4.28.0",
-    "analytics": "^0.7.21",
-    "bootstrap": "^5.0.2",
-    "boxible": "1.0.0",
-    "chokidar-cli": "^2.1.0",
-    "dayjs": "^1.10.6",
-    "eslint": "^7.29.0",
-    "eslint-plugin-react": "^7.24.0",
+    "@types/react-flatpickr": "^3.8.6",
+    "@types/react-router-dom": "^5.3.3",
+    "@typescript-eslint/eslint-plugin": "^5.18.0",
+    "@typescript-eslint/parser": "^5.18.0",
+    "analytics": "^0.8.1",
+    "bootstrap": "^5.1.3",
+    "boxible": "1.1.1",
+    "chokidar-cli": "^3.0.0",
+    "dayjs": "^1.11.0",
+    "eslint": "^8.12.0",
+    "eslint-plugin-react": "^7.29.4",
     "expect-playwright": "^0.8.0",
-    "faker": "^5.5.3",
     "formik": "^2.2.9",
     "loadjs": "^4.2.0",
     "lodash-es": "^4.17.21",
     "modeled-mobx": "^0.6.4",
-    "playwright": "^1.16.3",
-    "playwright-chromium": "^1.16.3",
+    "playwright": "^1.20.2",
+    "playwright-chromium": "^1.20.2",
     "qs": "^6.10.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-flatpickr": "^3.10.7",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-flatpickr": "^3.10.9",
     "react-popper": "^2.2.5",
-    "react-router-dom": "^5.2.0",
-    "react-select": "^5.1.0",
-    "rooks": "^5.7.3",
-    "sass": "^1.35.1",
-    "ts-node": "^10.0.0",
+    "react-router-dom": "^6.3.0",
+    "react-select": "^5.2.2",
+    "rooks": "^5.11.0",
+    "sass": "^1.49.11",
+    "ts-node": "^10.7.0",
     "url-parse": "^1.5.10",
-    "use-local-storage-state": "^15.0.0",
-    "wait-on": "^6.0.0",
+    "use-local-storage-state": "^17.0.0",
+    "wait-on": "^6.0.1",
     "whatwg-fetch": "^3.6.2",
-    "yup": "^0.32.9"
+    "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@types/react": "^17.0.34",
-    "@types/react-dom": "^17.0.11",
-    "@vitejs/plugin-react-refresh": "^1.3.3",
-    "typescript": "^4.3.4",
-    "vite": "^2.5.0"
+    "@faker-js/faker": "^6.1.2",
+    "@types/react-dom": "^17.0.14",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
+    "typescript": "^4.6.3",
+    "vite": "^2.9.1"
   }
 }
